+++ packages/miniapp-render/src/builtInComponents/index.js
@@ -32,6 +32,8 @@ import canvas from './canvas';
 import webView from './web-view';
 import livePlayer from './live-player';
 import livePusher from './live-pusher';
+import officialAccount from './official-account';
+import contactButton from './contact-button';
 
 // Sub components
 import movableView from './movable-view';
@@ -75,6 +77,8 @@ const components = [
   movableView,
   swiperItem,
   pickerViewColumn,
+  officialAccount,
+  contactButton
 ];
 
 const handlesMap = {