+++ packages/rax-miniapp-runtime-webpack-plugin/src/utils/wrapChunks.js
@@ -15,7 +15,7 @@ const matchFile = (fileName, ext) =>
 const FunctionPolyfill = 'Function||(Function=function(){return function(){return Symbol}}),void 0===Function.prototype.call&&(Function.prototype.call=function(n){(n=n||window).fn=this;const t=[...arguments].slice(1),o=n.fn(...t);return delete n.fn,o}),void 0===Function.prototype.apply&&(Function.prototype.apply=function(n){let t;return(n=n||window).fn=this,t=arguments[1]?n.fn(...arguments[1]):n.fn(),delete n.fn,t})';
 
 // Add content to chunks head and tail
-module.exports = function(compilation, chunks, pluginDir, target) {
+module.exports = function(compilation, chunks, target) {
   chunks.forEach((chunk) => {
     chunk.files.forEach((fileName) => {
       if (matchFile(fileName, 'js')) {